Include babel in install documentation
There have been some reports (#26) that the current bundle/install
process will error out if `babel` is not installed (either
globally or for the host project). I added a small step to
install `babel` and record it as a dependecy for the host
project.

On branch pull/document-babel-install
Changes to be committed:
	modified:   manual.md