Adjust behavior of Enter again
The lifting now only happens when the textblock is empty, and splitting
only happens when there is more content in the parent below the block.

This won't generalize quite as well to new wrapping node types, but
works okay with lists given the splitListItem command. We'll probably want
custom handlers for other list types.

Issue #88