Require registered values to have a name, allow computed ones
This makes it possible for subclassed node and mark types to selectively
override and delete their superclass's values, and for such values to
be omitted or changed based on the type they are registered on.

BREAKING:

The interface to SchemaItem.register changed, requiring a value name
as second argument. The various systems making use of the schema registry
now use this name, rather than a field in the values -- for commands and
input rules, it replaces spec.name, for dom parsers, the node name, and
for markdown parsers, the token name.