Fix menu icons visibility with pjax navigation
Pjax navigation allow to navigate without full page reload
by replacing page elements, usually the body, with dynamically
loaded content.
Prosemirror uses module constants to cache the SVG icons container
and to check if each icon has been already loaded or not.
This behaviour breaks when the pjax driver delete the body and replace
it with the new content: the cache objects are still set but the
SVG icons are not there anymore, hence they disappear from the menubar.